requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
numpy>=1.24.3
opencv-python>=4.8.1.78
Pillow>=10.1.0
scipy>=1.11.4
exifread>=3.0.0
pydantic>=2.12.0
typing-extensions>=4.15.0
